+++ src/test/mypy.sh
@@ -30,6 +30,7 @@ call_mypy ../hammer_config_test/test.py
 call_mypy ../hammer-vlsi/synthesis/mocksynth/__init__.py
 call_mypy ../hammer-vlsi/synthesis/vivado/__init__.py
 call_mypy ../hammer-vlsi/par/nop.py
+call_mypy ../hammer-vlsi/par/vivado/__init__.py
 
 # Plugins which may or may not exist
 if [ -f ../hammer-vlsi/synthesis/dc/__init__.py ]; then
